// models/DataProcessingAgreement.ts
/**
 * My Data My Consent - Developer API
 * Unleashing the power of data consent by establishing trust. The Platform Core Developer API defines a set of capabilities that can be used to request, issue, manage and update data, documents and credentials by organizations. The API can be used to request, manage and update Decentralised Identifiers, Financial Data, Health Data issue Documents, Credentials directly or using OpenID Connect flows, and verify Messages signed with DIDs and much more.
 *
 * OpenAPI spec version: v1
 * Contact: support@mydatamyconsent.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApplicationUser } from './ApplicationUser';
import { Organization } from './Organization';
import { HttpFile } from '../http/http';

export class DataProcessingAgreement {
    'createdBy'?: string;
    'createdAtUtc'?: Date;
    'updatedBy'?: string;
    'updatedAtUtc'?: Date;
    'createdByUser'?: ApplicationUser;
    'updatedByUser'?: ApplicationUser;
    'deletedBy'?: string;
    'deletedAtUtc'?: Date;
    'deletedByUser'?: ApplicationUser;
    'id'?: string;
    'organizationId'?: string;
    'version'?: string;
    'body'?: string;
    'attachmentUrl'?: string;
    'organization'?: Organization;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "createdBy",
            "baseName": "createdBy",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "createdAtUtc",
            "baseName": "createdAtUtc",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "updatedBy",
            "baseName": "updatedBy",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "updatedAtUtc",
            "baseName": "updatedAtUtc",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "createdByUser",
            "baseName": "createdByUser",
            "type": "ApplicationUser",
            "format": ""
        },
        {
            "name": "updatedByUser",
            "baseName": "updatedByUser",
            "type": "ApplicationUser",
            "format": ""
        },
        {
            "name": "deletedBy",
            "baseName": "deletedBy",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "deletedAtUtc",
            "baseName": "deletedAtUtc",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "deletedByUser",
            "baseName": "deletedByUser",
            "type": "ApplicationUser",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "organizationId",
            "baseName": "organizationId",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "version",
            "baseName": "version",
            "type": "string",
            "format": ""
        },
        {
            "name": "body",
            "baseName": "body",
            "type": "string",
            "format": ""
        },
        {
            "name": "attachmentUrl",
            "baseName": "attachmentUrl",
            "type": "string",
            "format": ""
        },
        {
            "name": "organization",
            "baseName": "organization",
            "type": "Organization",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DataProcessingAgreement.attributeTypeMap;
    }
    
    public constructor() {
    }
}